Improve project runtime test (safe/full) mode
