+++ microprofile.jdt/org.eclipse.lsp4mp.jdt.core/src/main/java/org/eclipse/lsp4mp/commons/runtime/converter/full/FullConverterRuntimeSupport.java
@@ -1,16 +1,16 @@
 /*******************************************************************************
-* Copyright (c) 2025 Red Hat Inc. and others.
-*
-* This program and the accompanying materials are made available under the
-* terms of the Eclipse Public License v. 2.0 which is available at
-* http://www.eclipse.org/legal/epl-2.0, or the Apache License, Version 2.0
-* which is available at https://www.apache.org/licenses/LICENSE-2.0.
-*
-* SPDX-License-Identifier: EPL-2.0 OR Apache-2.0
-*
-* Contributors:
-*     Red Hat Inc. - initial API and implementation
-*******************************************************************************/
+ * Copyright (c) 2025 Red Hat Inc. and others.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License v. 2.0 which is available at
+ * http://www.eclipse.org/legal/epl-2.0, or the Apache License, Version 2.0
+ * which is available at https://www.apache.org/licenses/LICENSE-2.0.
+ *
+ * SPDX-License-Identifier: EPL-2.0 OR Apache-2.0
+ *
+ * Contributors:
+ *     Red Hat Inc. - initial API and implementation
+ *******************************************************************************/
 package org.eclipse.lsp4mp.commons.runtime.converter.full;
 
 import java.lang.reflect.Method;
@@ -26,27 +26,61 @@ import org.eclipse.lsp4mp.commons.runtime.converter.AbstractConverterRuntimeSupp
 import org.eclipse.lsp4mp.commons.runtime.converter.ConverterValidator;
 
 /**
- * ConverterRuntimeSupport allows dynamic validation of string values against
- * Java types using MicroProfile Config converters from the project classpath.
+ * FULL execution mode implementation of
+ * {@link AbstractConverterRuntimeSupport}.
  *
  * <p>
- * This class uses reflection to avoid classloader issues and caches converters
- * per type for performance. Only converters discovered via
- * Config.getConverter(Class) are used.
+ * In FULL mode:
+ * </p>
+ * <ul>
+ * <li>The project classpath and project classes are accessible.</li>
+ * <li>Converters may be obtained from both standard and user-defined
+ * MicroProfile Config providers.</li>
+ * <li>Reflection is used to safely load and invoke
+ * {@code ConfigProviderResolver}, {@code ConfigBuilder}, and converters.</li>
+ * <li>Caches converters per type for performance.</li>
+ * </ul>
+ *
+ * <p>
+ * This class attempts to load a {@code ConfigProviderResolver} first from known
+ * SmallRye classes, then via {@link ServiceLoader}, and builds the
+ * {@code Config} using the project's runtime classloader.
  * </p>
  * 
- * @author Angelo ZERR
+ * <p>
+ * Once the {@code Config} is loaded, validation and conversion are delegated to
+ * {@link FullConverterValidator}.
+ * </p>
+ * 
+ * @author Angelo
  */
 public class FullConverterRuntimeSupport extends AbstractConverterRuntimeSupport<Object> {
 
 	private static final Logger LOGGER = Logger.getLogger(FullConverterRuntimeSupport.class.getName());
 
+	/** Default resolver class names to try first */
 	private static final String[] DEFAULT_RESOLVERS = { "io.smallrye.config.SmallRyeConfigProviderResolver" };
 
+	/**
+	 * Creates a new FULL runtime support instance for the given project.
+	 *
+	 * @param project the owning MicroProfile project runtime
+	 */
 	public FullConverterRuntimeSupport(MicroProfileProjectRuntime project) {
 		super(project, ExecutionMode.FULL);
 	}
 
+	/**
+	 * Loads the MicroProfile {@code Config} instance using the project classpath.
+	 *
+	 * <p>
+	 * Uses reflection to invoke {@code getBuilder()}, {@code forClassLoader()},
+	 * {@code addDiscoveredConverters()}, and {@code build()} on the first available
+	 * {@code ConfigProviderResolver}.
+	 * </p>
+	 *
+	 * @return the {@code Config} instance or {@code null} if none could be loaded
+	 */
 	@Override
 	protected Object loadConfig() {
 		Object resolver = loadConfigProviderResolverReflect();
@@ -61,7 +95,6 @@ public class FullConverterRuntimeSupport extends AbstractConverterRuntimeSupport
 				builder.getClass().getMethod("addDiscoveredConverters").invoke(builder);
 
 				return builder.getClass().getMethod("build").invoke(builder);
-
 			} catch (Throwable e) {
 				LOGGER.log(Level.INFO,
 						"Error creating MicroProfile Config via reflection from " + resolver.getClass().getName(), e);
@@ -70,13 +103,30 @@ public class FullConverterRuntimeSupport extends AbstractConverterRuntimeSupport
 		return null;
 	}
 
+	/**
+	 * Creates a new {@link FullConverterValidator} for the specified type using the
+	 * loaded {@code Config}.
+	 *
+	 * @param config the loaded MicroProfile Config instance
+	 * @param type   the target Java type to validate/convert
+	 * @return a {@link ConverterValidator} instance for FULL mode conversion
+	 */
 	@Override
 	protected ConverterValidator newConverter(Object config, Class<?> type) {
 		return new FullConverterValidator(config, type);
 	}
 
 	/**
-	 * Loads the first available ConfigProviderResolver.
+	 * Loads the first available {@code ConfigProviderResolver} from the project
+	 * classpath.
+	 *
+	 * <p>
+	 * First attempts to instantiate default known resolvers, then uses
+	 * {@link ServiceLoader}.
+	 * </p>
+	 *
+	 * @return the first available {@code ConfigProviderResolver} instance or
+	 *         {@code null} if none found
 	 */
 	private Object loadConfigProviderResolverReflect() {
 		ClassLoader runtimeCL = getProject().getRuntimeClassLoader();
@@ -117,5 +167,4 @@ public class FullConverterRuntimeSupport extends AbstractConverterRuntimeSupport
 
 		return resolvers.get(0);
 	}
-
 }
